# Conflicts:
#	Glidders/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
#	Glidders/Logs/AssetImportWorker0-prev.log
#	Glidders/Logs/AssetImportWorker0.log
#	Glidders/UserSettings/EditorUserSettings.asset
#	Glidders/obj/Debug/Assembly-CSharp.csproj.CoreCompileInputs.cache
#	Glidders/obj/Debug/Assembly-CSharp.csprojAssemblyReference.cache

+++ Glidders/Assets/Scripts/Field/FieldCore.cs
@@ -99,7 +99,7 @@ namespace Glidders
             public bool IsPassingGrid(FieldIndex fieldIndex)
             {
                 // グリッドの通行可否を返却する処理を記述してください。（trueなら通行可能）
-                return (fieldDeta[fieldIndex.row, fieldIndex.column] / 100 > (int)FieldCode.IMPENETABLE);
+                return ((int)(fieldDeta[fieldIndex.row, fieldIndex.column] / 100) > (int)FieldCode.IMPENETABLE);
             }
 
             public FieldIndex GetPlayerPosition(int playerNumber)
